+++ src/f_root.h
@@ -7,13 +7,13 @@
 
 /* f_root.c */
 struct f_root_params {
-  unsigned int verbose;
+  size_t verbose;
   double (*f) (double x, void *p);
   void *p;
   const gsl_root_fsolver_type *type;
   double eps_abs;
   double eps_rel;
-  unsigned int max_iter;
+  size_t max_iter;
 };
 
 int f_root (double x_min, double x_max, double *root, struct f_root_params *par);
